Add reusable NoOp sandbox implementation
